+++ CHANGELOG.md
@@ -31,6 +31,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Export is not working, when size of exported events did not match a multiple of buffer size
+
 ### Security
 
 ## 2.0.3