+++ CHANGELOG.md
@@ -21,10 +21,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Internal: changed gradle lock mode
 - Force remove of account data on logout
 - Improvements in ExportRoom: parallel decryption, collect decryption errors
-- Move gradle locks to CI
+- Move Gradle locks to CI
 - Restrict ability to change room history to WORLD_READABLE when the group is encrypted
 - Fixed typo in RoomEncryptionEnabledViewModel
-- Upgrade gradle to 8.9
+- Upgrade Gradle to 8.9
+- Fixed Java toolchain configuration in emoji subproject
 
 ### Deprecated